requirements.txt
Django==3.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-redis==5.4.0
requests==2.32.3
pandas==2.2.2
shapely==2.0.4
pyproj==3.6.1
matplotlib==3.9.2
Pillow==10.4.0
python-dotenv==1.0.1
redis==5.0.1
httpx==0.25.2
psycopg2-binary==2.9.9
django-environ==0.11.2
geopy==2.4.1